requirements.txt
aiohttp==3.7.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
django-user-agents==0.4.0
djangorestframework==3.12.2
geoip2==4.1.0
gunicorn==20.0.4
idna==2.10
maxminddb==2.0.3
multidict==5.0.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
urllib3==1.26.2
user-agents==2.2.0
whitenoise==5.2.0
yarl==1.6.3
